Split tree_utils from utils
