+++ docs/api/tree_utils.md
@@ -1,0 +1,1 @@
+::: posteriors.tree_utils
