+++ posteriors/__init__.py
@@ -16,19 +16,20 @@ from posteriors.utils import diag_ggn
 from posteriors.utils import cg
 from posteriors.utils import diag_normal_log_prob
 from posteriors.utils import diag_normal_sample
-from posteriors.utils import tree_size
-from posteriors.utils import tree_extract
-from posteriors.utils import tree_insert
-from posteriors.utils import tree_insert_
-from posteriors.utils import extract_requires_grad
-from posteriors.utils import insert_requires_grad
-from posteriors.utils import insert_requires_grad_
-from posteriors.utils import extract_requires_grad_and_func
-from posteriors.utils import inplacify
-from posteriors.utils import tree_map_inplacify_
-from posteriors.utils import flexi_tree_map
 from posteriors.utils import per_samplify
 from posteriors.utils import is_scalar
+
+from posteriors.tree_utils import tree_size
+from posteriors.tree_utils import tree_extract
+from posteriors.tree_utils import tree_insert
+from posteriors.tree_utils import tree_insert_
+from posteriors.tree_utils import extract_requires_grad
+from posteriors.tree_utils import insert_requires_grad
+from posteriors.tree_utils import insert_requires_grad_
+from posteriors.tree_utils import extract_requires_grad_and_func
+from posteriors.tree_utils import inplacify
+from posteriors.tree_utils import tree_map_inplacify_
+from posteriors.tree_utils import flexi_tree_map
 
 import logging
