+++ posteriors/sgmcmc/sghmc.py
@@ -6,7 +6,8 @@ from optree import tree_map
 from dataclasses import dataclass
 
 from posteriors.types import TensorTree, Transform, LogProbFn, TransformState
-from posteriors.utils import flexi_tree_map, is_scalar, CatchAuxError
+from posteriors.tree_utils import flexi_tree_map
+from posteriors.utils import is_scalar, CatchAuxError
 
 
 def build(
